# Optimize BN254 Pairing