--- a/fuzz/no_gpu_fuzz/fuzz_targets/deserialize_stark_proof.rs
+++ b/fuzz/no_gpu_fuzz/fuzz_targets/deserialize_stark_proof.rs
@@ -1,12 +1,9 @@
 #![no_main]
-use libfuzzer_sys::fuzz_target;
-use stark_platinum_prover::proof::stark::StarkProof;
 use lambdaworks_math::field::fields::fft_friendly::stark_252_prime_field::Stark252PrimeField;
 use lambdaworks_math::traits::Deserializable;
-
+use libfuzzer_sys::fuzz_target;
+use stark_platinum_prover::proof::stark::StarkProof;
 
 fuzz_target!(|data: Vec<u8>| {
-    
     let _proof = StarkProof::<Stark252PrimeField>::deserialize(&data);
-
 });