--- a/fuzz/no_gpu_fuzz/fuzz_targets/field/stark_field_addition.rs
+++ b/fuzz/no_gpu_fuzz/fuzz_targets/field/stark_field_addition.rs
@@ -1,10 +1,9 @@
 #![no_main]
 
-use libfuzzer_sys::fuzz_target;
 use lambdaworks_math::field::{
-    element::FieldElement, 
-    fields::fft_friendly::stark_252_prime_field::Stark252PrimeField,
+    element::FieldElement, fields::fft_friendly::stark_252_prime_field::Stark252PrimeField,
 };
+use libfuzzer_sys::fuzz_target;
 
 use lambdaworks_math::traits::ByteConversion;
 
@@ -12,7 +11,7 @@ type FE = FieldElement<Stark252PrimeField>;
 
 use ibig::{modular::ModuloRing, UBig};
 
-fuzz_target!(|bytes: ([u8;32], [u8;32])| {
+fuzz_target!(|bytes: ([u8; 32], [u8; 32])| {
     let (bytes_a, bytes_b) = bytes;
     let a = FE::from_bytes_be(&bytes_a).unwrap();
     let b = FE::from_bytes_be(&bytes_b).unwrap();
@@ -23,8 +22,11 @@ fuzz_target!(|bytes: ([u8;32], [u8;32])| {
     let c = a + &b;
     let c_hex = c.representative().to_string()[2..].to_string();
 
-    let prime = 
-    UBig::from_str_radix("800000000000011000000000000000000000000000000000000000000000001", 16).unwrap();
+    let prime = UBig::from_str_radix(
+        "800000000000011000000000000000000000000000000000000000000000001",
+        16,
+    )
+    .unwrap();
     let cairo_ring = ModuloRing::new(&prime);
 
     let a_ring = cairo_ring.from(&UBig::from_str_radix(&a_hex, 16).unwrap());
@@ -34,4 +36,3 @@ fuzz_target!(|bytes: ([u8;32], [u8;32])| {
 
     assert_eq!(expected_c_hex, c_hex);
 });
-