--- a/math/benches/criterion_elliptic_curve.rs
+++ b/math/benches/criterion_elliptic_curve.rs
@@ -9,7 +9,7 @@ use elliptic_curves::{
 
 criterion_group!(
     name = elliptic_curve_benches;
-    config = Criterion::default().with_profiler(PProfProfiler::new(100, Output::Flamegraph(None)));
+    config = Criterion::default();
     targets =  bn_254_elliptic_curve_benchmarks, 
 );