# BN254 optimization 

## Description

This PR improves the paring performance by using optimized operations used in key functions.
Also it incorporates benches for the BN254.

There is still room for improvement optimizng the Miller loop and the operations for Fp6 and Fp12


Optimized algorithms:

- ``Line`` 
- ``final exponentiation`` 
    - ```cyclotomic_square```
  - ``cyclotomic_pow``

- ``Fp2 operations``


## Type of change
- [ ] Optimization

## Benches
- Actual
  ```Ate pairing``` : 1.6898 ms 
  ```Miller loop``` : 540.37 µs
  ```Final exponentiation```: 771.49 µs

- New version
  ```Ate pairing``` :939.71 µs 
  ```Miller loop``` : 375.96 µs
  ```Final exponentiation```: 347.71 µs